news_api/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
langchain
langgraph
langchain-google-vertexai
google-cloud-aiplatform